# Vector caching.