As in the title. Benchmarks would be nice if anyone has time for doing them.