@Starbuck5 2025-02-03
> Benchmarks would be nice if anyone has time for doing them.

I think it is the PR author's job to justify why the change should be added. Writing the code is often not the hard part, it's doing the research, doing experimentation, making sure others understand. From what I've seen that still needs to happen, so I'm marking this PR as a draft.