@gresm 2025-02-03
@Starbuck5
---
Using this timing script:
```py
import time

def test(rounds):
    import pygame
    start = time.perf_counter()
    for _ in range(rounds):
        vec = pygame.Vector3(1, 2, 3)
        vec2 = vec.copy()
    end = time.perf_counter()
    print(f"Total time: {end - start}, on average: {(end - start) / rounds}")
```

## For ``test(10**7)`` and ``VECTOR_CACHE_SIZE==100000``
Before:
```
Total time: 1.37842830199952, on average: 1.37842830199952e-07
```
After:
```
Total time: 1.2690909360026126, on average: 1.2690909360026126e-07
```
## For ``test(10**7)`` and ``VECTOR_CACHE_SIZE==10000``
Before:
```
Total time: 1.3938012459984748, on average: 1.3938012459984747e-07
```
After:
```
Total time: 1.2679740370003856, on average: 1.2679740370003856e-07
```

## For ``test(10**7)`` and ``VECTOR_CACHE_SIZE==1000``
Before:
```
Total time: 1.392446205001761, on average: 1.392446205001761e-07
```
After:
```
Total time: 1.2817346929987252, on average: 1.2817346929987253e-07
```

## For ``test(10**7)`` and ``VECTOR_CACHE_SIZE==100``
Before:
```
Total time: 1.3991181530000176, on average: 1.3991181530000176e-07
```
After:
```
Total time: 1.2715259210017393, on average: 1.2715259210017393e-07
```

## For ``test(10**7)`` and ``VECTOR_CACHE_SIZE==10``
Before:
```
Total time: 1.3997546829996281, on average: 1.399754682999628e-07
```
After:
```
Total time: 1.2715557740011718, on average: 1.271555774001172e-07
```

## For ``test(10**7)`` and ``VECTOR_CACHE_SIZE==1``
Before:
```
Total time: 1.4705402879990288, on average: 1.4705402879990288e-07
```
After:
```
Total time: 1.2530362709985639, on average: 1.253036270998564e-07
```
## Conclusion
There's a difference between them, but I probably messed up benchmarks a bit. Maybe instancing multiple vectors at the same time would make more sense.